backend/requirements.txt
fastapi
uvicorn[standard]
redis
discord.py
pydantic
pydantic-settings
python-dotenv
websockets
pyjwt
